Verilog: big_endian -> increasing
This changes the terminology for increasing vs. vector index ranges in the
implementation of the Verilog frontend (the change is not user-visible).

The term "big endian" is commonly used for _byte_ orderings.  By contrast,
the use here is for the numbering of _bits_ in a vector.  The SystemVerilog
standard does not introduce terminology for this, but there is a system
named function $increment that returns -1 or 1, depending on the numbering
direction.