SVA/LTL property instrumentation
This adds --buechi, which triggers an instrumentation pass that turns LTL
and select SVA properties into a Buechi automaton.  The automaton is then
added to the model, and the property is replaced by the Buechi acceptance
condition.

The translation is done via ltl2tgba.

This enables checking arbitrary LTL properties via the BDD engine.