SystemVerilog: create identifier expressions in scanner
The verilog_identifer expressions are now created in the scanner, not the
parser.

The benefit of doing so is that the lookup needed in the scanner to
determine the token type does not need to be repeated in the parser to
identify the scope in which the identifier was defined in.