SVA: and/or as a sequence also need to work with parentheses
The SVA operators 'and' and 'or' can be used to form SVA sequences or SVA
properties, depending on the types of the operands.  Furthermore, any
sequence expression is also a property expression.

This results in a conflict in the SystemVerilog grammar, as the context is
insufficient to determine whether an expression is a sequence_expr or a
property_expr.

This resolves the conflict by replicating the sequence_expr rules into
property_expr, and allowing property_expr to be used where sequence_expr
would be required.

To detect malformed input programs that use a property where a sequence is
required, a check is added to the type checker.