SVA: sequence property expressions
SVA Sequences can be used as properties, in one of three ways:

1. weak(s)

2. strong(s)

3. s

The third form is interpreted as strong when used in a cover statement, and
as weak in assert/assume. The SVA not operator flips this around.

This adds a new expression, sva_sequence_property_exprt, to denote this
conversion.