SVA: default sequence semantics
SVA sequences can have weak or strong semantics.  1800-2017 16.12.2 Sequence
property requires that sequences in properties that are not explicitly
marked as strong or weak are interpreted as weak when used in assert
property or assume property, and strong otherwise.

The Verilog type checker now emits corresponding expressions
sva_implict_weak and sva_implicit_strong for uses of SVA sequences in
properties.