BMC: reconsider lasso-shaped encodings
This replaces the lasso-shaped encodings in the word-level BMC property
encoding.  The lasso and straigth-line encodings are now built as separate
obligations.  In the lasso encoding, all operators follow the same lasso
shape.  This avoids spurious traces where the eventually operator follows
the lasso but a safety part of the property does not.