LTL/SVA->Buechi: reject error traces early
The Buechi automaton accepts counterexamples to the original property.  The
Buechi acceptance condition requires a lasso where the loop contains an
accepting state.  A proof of that requires satisfying the looping condition,
which may require a very long counterexample.

This adds a shortcut using a sufficient condition for a counterexample: we
can reject the property when we reach a state in which the Buechi automaton
transitions into an accepting state that has an unconditional self-loop on
it.  Such traces can always be extended to a lasso as above.