Buechi: identify special cases where reachability is sufficient
This adds treatment for the special case where all nonaccepting states of
the Buechi automaton have an unconditional self-loop.

In this case, when these nonaccepting states is reached, the trace cannot be
extended to an accepting trace, and will hence be rejected.

Hence, in this case, the Buechi acceptance condition is unnecessary, and a
simple reachability property is sufficient.

Among the 74 tests in regression/ebmc-spot/sva-buechi, only 3 require the
Büchi acceptance condition with this change applied.