Verilog: assignment conversion
The conversion done to the RHS of an assignment differs from the downwards
propagation done when adjusting the type of the operand of expressions.

This introduces a new method assignment_conversion, different from
propagate_type, to implement these differences.

Three tests are changed to restore the previous circuit, as the circuit
without the downward propagation pass is too large for BDDs.