Verilog: move assignment conversion code
The code for assignment conversions (1800-2017 10.8) is required for
converting the arguments of function calls, and hence, when type checking
expressions.

This moves the code for assignment conversions from verilog_typecheck.cpp to
verilog_typecheck_expr.cpp, and replaces the use of propagate_type by
assignment_conversion.  The propagate_type method is then no longer needed.